+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [unreleased]
+### Added
+ - Show locale date as tooltip for relative date time information
+
 ## [8.5.0] - 2025-10-03
 ### Fixed
  - Archived polls could not get entered anymore